+++ docassemble_base/docassemble/base/error.py
@@ -23,7 +23,7 @@ class DAValidationError(Exception):
 
 class CodeExecute(Exception):
     def __init__(self, compute, question):
-        if type(compute) is list:
+        if isinstance(compute, list):
             self.compute = "\n".join(compute)
         else:
             self.compute = compute
@@ -55,7 +55,7 @@ class ForcedNameError(NameError):
         for var_name in ('x', 'i', 'j', 'k', 'l', 'm', 'n'):
             if var_name in the_user_dict:
                 the_context[var_name] = the_user_dict[var_name]
-        if type(the_args[0]) is dict:
+        if isinstance(the_args[0], dict):
             if 'action' in the_args[0] and (len(the_args[0]) == 1 or 'arguments' in the_args[0]):
                 self.name = the_args[0]['action']
                 self.arguments = the_args[0].get('arguments', dict())
@@ -70,16 +70,18 @@ class ForcedNameError(NameError):
         self.next_action = list()
         while len(the_args):
             arg = the_args.pop(0)
-            if type(arg) is dict:
+            if isinstance(arg, dict):
                 if (len(arg.keys()) == 2 and 'action' in arg and 'arguments' in arg) or (len(arg.keys()) == 1 and 'action' in arg):
                     self.set_action(arg)
                 elif len(arg) == 1 and ('undefine' in arg or 'invalidate' in arg or 'recompute' in arg or 'set' in arg or 'follow up' in arg):
                     if 'set' in arg:
-                        if type(arg['set']) is not list:
+                        if isinstance(arg['set'], dict):
+                            arg['set'] = [arg['set']]
+                        if not isinstance(arg['set'], list):
                             raise DAError("force_ask: the set statement must refer to a list.")
                         clean_list = []
                         for the_dict in arg['set']:
-                            if type(the_dict) is not dict:
+                            if not isinstance(the_dict, dict):
                                 raise DAError("force_ask: a set command must refer to a list of dicts.")
                             for the_var, the_val in the_dict.items():
                                 if not isinstance(the_var, str):
@@ -90,27 +92,31 @@ class ForcedNameError(NameError):
                             clean_list.append([the_var_stripped, the_val])
                         self.set_action(dict(action='_da_set', arguments=dict(variables=clean_list), context=the_context))
                     if 'follow up' in arg:
-                        if type(arg['follow up']) is not list:
+                        if isinstance(arg['follow up'], str):
+                            arg[command] = [arg['follow up']]
+                        if not isinstance(arg['follow up'], list):
                             raise DAError("force_ask: the follow up statement must refer to a list.")
                         for var in arg['follow up']:
                             if not isinstance(var, str):
                                 raise DAError("force_ask: invalid variable name " + repr(var) + " in follow up.")
                             var_saveas = var.strip()
                             if invalid_variable_name(var_saveas):
-                                raise DAError("force_ask: missing or invalid variable name " + repr(var_saveas) + " .  " + repr(data))
+                                raise DAError("force_ask: missing or invalid variable name " + repr(var_saveas) + ".")
                             self.set_action(dict(action=var, arguments=dict(), context=the_context))
                     for command in ('undefine', 'invalidate', 'recompute'):
                         if command not in arg:
                             continue
-                        if type(arg[command]) is not list:
-                            raise DAError("force_ask: the " + command + " statement must refer to a list.  " + repr(data))
+                        if isinstance(arg[command], str):
+                            arg[command] = [arg[command]]
+                        if not isinstance(arg[command], list):
+                            raise DAError("force_ask: the " + command + " statement must refer to a list.  ")
                         clean_list = []
                         for undef_var in arg[command]:
                             if not isinstance(undef_var, str):
-                                raise DAError("force_ask: invalid variable name " + repr(undef_var) + " in " + command + ".  " + repr(data))
+                                raise DAError("force_ask: invalid variable name " + repr(undef_var) + " in " + command + ".")
                             undef_saveas = undef_var.strip()
                             if invalid_variable_name(undef_saveas):
-                                raise DAError("force_ask: missing or invalid variable name " + repr(undef_saveas) + " .  " + repr(data))
+                                raise DAError("force_ask: missing or invalid variable name " + repr(undef_saveas) + ".")
                             clean_list.append(undef_saveas)
                         if command == 'invalidate':
                             self.next_action.append(dict(action='_da_invalidate', arguments=dict(variables=clean_list), context=the_context))